+++ src/Applet/Generate/appletGenerateCommand.ts
@@ -122,9 +122,11 @@ export const appletGenerate = createCommandDefinition({
 
 		// Git support select
 		let git: GitOptions | undefined = options.git as GitOptions | undefined;
-		let gitFound = await executeChildProcess('git --version', 'Git not found on this machine', false);
+		let gitFound = await executeChildProcess('git --version', false).catch((err: string) => {
+			console.error(`Git not found on this machine: ${err}`);
+		});
 		// Skip prompt if git was not found
-		if (git === undefined && gitFound.includes('git version')) {
+		if (git === undefined && gitFound?.includes('git version')) {
 			const response = await prompts({
 				type: 'select',
 				name: 'git',
@@ -434,17 +436,17 @@ always-auth=true
 
 const initGitRepository = (directoryPath: string): void => {
 	const absolutePath = path.resolve(directoryPath);
-	executeChildProcess(`git init "${absolutePath}"`, 'Git repository initialization failed', true);
+	executeChildProcess(`git init "${absolutePath}"`, true).catch((err: string) => {
+		console.error(`Git repository initialization failed: ${err}`);
+	});
 };
 
-const executeChildProcess = (command: string, errorMessage: string, verbose: boolean): Promise<string> => {
+const executeChildProcess = (command: string, verbose: boolean): Promise<string> => {
 	return new Promise((resolve, reject) => {
 		child_process.exec(command, (error, stdout, stderr) => {
 			if (error) {
-				console.error(`${errorMessage}: ${error.message}`);
 				reject(error.message);
 			} else if (stderr) {
-				console.error(`Operation failed: ${stderr}`);
 				reject(stderr);
 			} else {
 				if (verbose) {