+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Updated `tsconfig.js` rules to match current framework features
 - Updated `package.json` minimal supported node/npm engine versions
+- Updated dependencies
 - Updated documentation
 
 ## [1.10.0] - 2025-03-24