+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Updated `tsconfig.js` rules to match current framework features
+- Updated `package.json` minimal supported node/npm engine versions
+- Updated dependencies
+- Updated documentation
+
 ## [1.10.0] - 2025-03-24
 ### Added
 - New feature to init generated applet as git repository (through wizard or `--git yes`) when `git` command is present on machine