+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added option to support Rspack for generated applets
+- Removed support for Esbuild
+
 ## [2.0.0] - 2025-03-28
 ### Changed
 - Updated `tsconfig.js` rules to match current framework features