+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - New command `sos custom-script generate` to generate local repository with Custom Script boilerplate
 
+### Fixed
+- Argument --git is now optional
+- Git detection on windows
+- Include applet generator template files
+
 ## [2.1.0] - 2025-04-08
 ### Added
 - Added option to support Rspack for generated applets