+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - New command `sos custom-script generate` to generate local repository with Custom Script boilerplate
 
+### Fixed
+- Fixed issue on `sos applet generate` when optional `--git` property was required
+- Fixed issue on `sos applet generate` with Git support detection on Windows 
+- Fixed missing template files for `sos applet generate` command
+
 ## [2.1.0] - 2025-04-08
 ### Added
 - Added option to support Rspack for generated applets